requirements.txt
kafka-python==2.0.2
Faker==28.4.1
setuptools
apache-flink==1.20.0
